Updated bug triggering command to recreate the absolute path programatically
Java doesn't recognize "~" natively and thus cannot expand it into the location of the home directory even though on UNIX-like system this directory has a predefined pattern. We never encountered this issue before because we always knew our own user credentials, but now that others might run this command it is no longer the case.